Improve FTS query performance dramatically
Something about query plans on a virtual table (like FTS) unable to be
optmized, resulting in use of a slow correlated subquery.
Using IN instead of EXISTS improves query time by up to ~700-1400x
when the search is on a large FTS index for a large set of main entities.